+++ package.json
@@ -10,25 +10,24 @@
     "start": "forever stopall && forever start index.js",
     "deploy": "node deploy.js"
   },
-  "author": "Ryan Lewis <ryan@ryanhlewis.com>",
+  "author": "Ryan Lewis <ryan@ryanlewis.dev>",
   "license": "MIT",
   "dependencies": {
-    "@hapi/boom": "9.1.2",
+    "@hapi/boom": "9.1.3",
     "@hapi/catbox": "11.1.1",
     "@hapi/catbox-redis": "6.0.2",
     "@hapi/cookie": "11.0.2",
-    "@hapi/hapi": "20.1.4",
+    "@hapi/hapi": "20.1.5",
     "@hapi/inert": "6.0.3",
     "archiver": "5.3.0",
-    "array-shuffle": "2.0.0",
-    "aws-sdk": "2.934.0",
+    "array-shuffle": "3.0.0",
+    "aws-sdk": "2.954.0",
     "bcryptjs": "2.4.3",
     "bluebird": "3.7.2",
     "cli-spinner": "0.2.10",
     "connected-react-router": "6.9.1",
     "forever": "4.0.1",
     "glob": "7.1.7",
-    "history": "4.10.1",
     "lodash": "4.17.21",
     "moment": "2.29.1",
     "mysql2": "2.2.5",
@@ -42,7 +41,7 @@
     "redux-thunk": "2.3.0",
     "request": "2.88.0",
     "scp2": "0.5.0",
-    "sequelize": "6.6.2",
+    "sequelize": "6.6.5",
     "serialize": "0.1.3",
     "ssh-exec": "2.0.0",
     "stream-buffers": "3.0.2",
@@ -50,26 +49,27 @@
     "uuid": "8.3.2"
   },
   "devDependencies": {
-    "@babel/core": "7.14.6",
+    "@babel/core": "7.14.8",
     "@babel/plugin-proposal-class-properties": "7.14.5",
     "@babel/plugin-proposal-object-rest-spread": "7.14.7",
-    "@babel/preset-env": "7.14.7",
+    "@babel/preset-env": "7.14.8",
     "@babel/preset-react": "7.14.5",
     "babel-loader": "8.2.2",
     "config": "3.3.6",
-    "css-loader": "5.2.6",
-    "css-minimizer-webpack-plugin": "3.0.1",
-    "eslint": "7.13.0",
+    "css-loader": "6.2.0",
+    "css-minimizer-webpack-plugin": "3.0.2",
+    "eslint": "7.31.0",
+    "history": "4.10.1",
     "less": "4.1.1",
-    "less-loader": "10.0.0",
+    "less-loader": "10.0.1",
     "loglevel": "1.7.1",
-    "mini-css-extract-plugin": "1.6.0",
+    "mini-css-extract-plugin": "2.1.0",
     "redux-devtools-extension": "2.13.9",
     "standard": "16.0.3",
-    "style-loader": "3.0.0",
+    "style-loader": "3.2.1",
     "svg-inline-loader": "0.8.2",
-    "terser-webpack-plugin": "5.1.3",
-    "webpack": "5.40.0",
+    "terser-webpack-plugin": "5.1.4",
+    "webpack": "5.47.0",
     "webpack-cli": "4.7.2"
   }
 }
