+++ scripts/module_03/create-ami.js
