+++ scripts/module_03/create-ec2-instance.js
