+++ scripts/module_03/manage-ec2-instance.js
