+++ scripts/module_04/create-auto-scaling.js
