+++ scripts/module_04/create-launch-configuration.js
