+++ scripts/module_05/create-ec2-instance.js
