+++ scripts/module_05/helpers.js
@@ -1,115 +1,55 @@
-const AWS = require('aws-sdk')
-const config = require('config')
+const glob = require('glob')
+const fs = require('fs')
 
-const assets = require('../../util/assets')
+function getPublicFiles () {
+  return new Promise((resolve, reject) => {
+    glob('../../public/**/*.*', (err, files) => {
+      if (err) reject(err)
+      else {
+        const filePromises = files.map((file) => {
+          return new Promise((resolve, reject) => {
+            fs.readFile(file, (err, data) => {
+              if (err) reject(err)
+              else resolve(data)
+            })
+          })
+        })
 
-const awsRegion = config.get('aws.region')
-AWS.config.update({ region: awsRegion })
-
-const hamsterData = [
-  {
-    id: 1,
-    name: 'Zepto',
-    type: 'Speedball',
-    src: assets.hamster1,
-    results: []
-  }, {
-    id: 2,
-    name: 'Milkshake',
-    type: 'Speedball',
-    src: assets.hamster2,
-    results: []
-  }, {
-    id: 3,
-    name: 'Fievel',
-    type: 'Tiny Terror',
-    src: assets.hamster3,
-    results: []
-  }, {
-    id: 4,
-    name: 'Baby Ham',
-    type: 'Roller',
-    src: assets.hamster4,
-    results: []
-  }, {
-    id: 5,
-    name: 'Tater',
-    type: 'Stealth',
-    src: assets.hamster5,
-    results: []
-  }, {
-    id: 6,
-    name: 'Peter Pan',
-    type: 'ZigZagger',
-    src: assets.hamster6,
-    results: []
-  }
-]
-const raceData = [
-  {
-    id: 1,
-    venue: 'Petco 2000',
-    city: 'Seattle, WA',
-    date: '04/29/17',
-    results: []
-  }, {
-    id: 2,
-    venue: 'Triscuit Circuit 4700',
-    city: 'Daytona Beach, FL',
-    date: '09/21/17',
-    results: []
-  }, {
-    id: 3,
-    venue: 'Kraft 35',
-    city: 'Tokyo, Japan',
-    date: '07/14/17',
-    results: []
-  }
-]
-
-async function getHamsterData () {
-  return hamsterData
+        Promise.all(filePromises)
+          .then((fileContents) => {
+            return fileContents.map((contents, i) => {
+              return {
+                contents,
+                name: files[i].replace('../../public/', '')
+              }
+            })
+          })
+          .then(resolve)
+          .catch(reject)
+      }
+    })
+  })
 }
 
-async function getRaceData () {
-  return raceData
-}
-
-async function createSecurityGroup (sgName, port) {
-  const ec2 = new AWS.EC2()
-  const params = {
-    Description: sgName,
-    GroupName: sgName
+function getContentType (filename) {
+  if (filename.match(/\.html/)) {
+    return 'text/html'
   }
-
-  try {
-    const data = await ec2.createSecurityGroup(params).promise()
-
-    const ruleParams = {
-      GroupId: data.GroupId,
-      IpPermissions: [
-        {
-          IpProtocol: 'tcp',
-          FromPort: port,
-          ToPort: port,
-          IpRanges: [
-            {
-              CidrIp: '0.0.0.0/0'
-            }
-          ]
-        }
-      ]
-    }
-
-    await ec2.authorizeSecurityGroupIngress(ruleParams).promise()
-    return data.GroupId
-  } catch (err) {
-    throw new Error(`Error creating Security Group: ${err}`)
+  if (filename.match(/\.png/)) {
+    return 'image/png'
+  }
+  if (filename.match(/\.jpg/)) {
+    return 'image/jpeg'
+  }
+  if (filename.match(/\.js/)) {
+    return 'text/javascript'
+  }
+  if (filename.match(/\.css/)) {
+    return 'text/css'
   }
 }
 
 module.exports = {
-  getHamsterData,
-  getRaceData,
-  createSecurityGroup
+  getPublicFiles,
+  getContentType
 }
