+++ scripts/module_05/upload-s3-objects.js
