+++ scripts/module_06/create-dynamo-table.js
