+++ scripts/module_06/create-rds-db.js
