+++ scripts/module_06/create-redis-cluster.js
