+++ scripts/module_06/populate-dynamo-table.js
