+++ scripts/module_08/configure-s3-website.js
