+++ scripts/module_08/create-cloudfront-distribution.js
