+++ scripts/module_09/create-kinesis-stream.js
