+++ scripts/module_09/create-sqs-queue.js
