+++ scripts/module_09/lambda-kinesis-consumer/index.js
