+++ scripts/module_10/create-cloudwatch-alarm.js
